[docs] Lazy load demo toolbar
